Add support for remote sudo and prefs
To use cdist in Amazon ec2 with the official image from Ubuntu, I need to have the remote user run as "ubuntu" and not "root". To support this, the remote commands are wrapped in "sudo" and a "sudo chown -R ..." is needed. Also, the file "bin/cdist-my-config" is an additional config file that is sourced in scripts that have $__cdist_target_host set and sets __cdist_remote_user and __cdist_remote_sudo. This is similar to having host-specific settings in .ssh/config.