Bootstrap creates new repository, not a "bare" one
"Bare" repositories are a special thing in Git, specifically, they're repositories with no working tree, instead just having the contents of `.git`.

The bootstrap document says it's creating a "bare" repository, but it appears to just be creating a plain new Git repository. This commit changes the naming to be "new", to remove any confusion for people who know about bare Git repositories.