crosslink html man pages
This is using sed --in-place, which might not be available in
all sed versions. If that's a concern, then please replace with
awk script or move to build-helper and use temporary files.

Also the regex is a heuristic. It works for our man pages here
but it might have false positive matches in the future.